ci(scripts): extract shared release-script lib
Both release scripts duplicated changelog parsing, process exec, the
workspace package map, and the released-packages detection; comment-
released-prs also imported parsers straight from the enrich script. Pull
the shared pieces into scripts/release/lib (changelog, exec, workspace,
released-packages), with shared changelog.fixtures for the tests, so each
entry point keeps only its own logic and depends on the lib, not on its
sibling.
